% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Creates a Bootstrap Distribution for a given statistic using an given empirical distribution}
\usage{
bootstrap(emp_dist, B = 5000, stat, quantile = NULL)
}
\arguments{
\item{emp_dist}{Empirical Distribution to bootstrap from}

\item{B}{Number of bootstrap iterations to perform. Default = 5000}

\item{stat}{Statistic to create bootstrap distribution for. Options: mean, median, sd, iqr, quantile}

\item{quantile}{Quantile to calculate between 0 and 1 if stat = quantile}
}
\value{
A vector of size B containing bootstrap distribution for given statistic
}
\description{
Creates a Bootstrap Distribution for a given statistic using an given empirical distribution
}
