% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_loc_stat.R
\name{bs_loc_stat}
\alias{bs_loc_stat}
\title{Computes estimated bootstrap mean, bootstrap sd, percentile interval, and pivotal interval for a location-based statistic (mean, median, quantile)}
\usage{
bs_loc_stat(emp_dist, B = 5000, stat, quantile = NULL, alpha = 0.05)
}
\arguments{
\item{emp_dist}{Empirical distribution to bootstrap}

\item{B}{Number of bootstrap iterations to perform}

\item{stat}{Location-based statistic to use. Options: mean, median, or quantile}

\item{quantile}{Quantile to calculate between 0 and 1 if stat = quantile}

\item{alpha}{Alpha level for CI's. Default alpha = 0.05 for 95\% conf. Conf = (1 - alpha)*100\%}
}
\value{
A list containing the boostrap mean, bootsrap sd, percentile interval, and pivotal interval
}
\description{
Computes estimated bootstrap mean, bootstrap sd, percentile interval, and pivotal interval for a location-based statistic (mean, median, quantile)
}
