% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_plot.R
\name{bs_plot}
\alias{bs_plot}
\title{Plots the bootstrap distribution}
\usage{
bs_plot(boot_dist, stat)
}
\arguments{
\item{boot_dist}{Bootstrap Distribution}

\item{stat}{A string corresponding to the statistic of boostrap distribution}
}
\value{
A plot of the boostrap distribution
}
\description{
Plots the bootstrap distribution
}
