% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_scale_stat.R
\name{bs_scale_stat}
\alias{bs_scale_stat}
\title{Computes estimated bootstrap mean, bootstrap sd, percentile interval, and pivotal interval for a scale-based statistic (sd, iqr)}
\usage{
bs_scale_stat(emp_dist, B = 5000, stat, alpha = 0.05)
}
\arguments{
\item{emp_dist}{Empirical distribution to bootstrap}

\item{B}{Number of bootstrap iterations to perform}

\item{stat}{Scale-based statistic to use. Options: sd, iqr}

\item{alpha}{Alpha level for CI's. Default alpha = 0.05 for 95\% conf. Conf = (1 - alpha)*100\%}
}
\value{
A list containing the boostrap mean, bootsrap sd, percentile interval, and pivotal interval
}
\description{
Computes estimated bootstrap mean, bootstrap sd, percentile interval, and pivotal interval for a scale-based statistic (sd, iqr)
}
