% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_summarize.R
\name{bs_summarize}
\alias{bs_summarize}
\title{Prints summary statistics and plot of bootstrap distribution}
\usage{
bs_summarize(emp_dist, B = 5000, stat, quantile = NULL, alpha = 0.05)
}
\arguments{
\item{emp_dist}{Empirical distribution to bootstrap}

\item{B}{Number of bootstrap iterations to perform}

\item{stat}{Measure of center to use as statistic: mean, median, or quantile}

\item{quantile}{Quantile to calculate between 0 and 1 if stat = quantile}

\item{alpha}{Alpha level for CI's: Conf = (1 - alpha)}
}
\value{
A table and plot
}
\description{
Prints summary statistics and plot of bootstrap distribution
}
