% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{calc_theta}
\alias{calc_theta}
\title{Calculates a specified statistic on a bootstrap sample}
\usage{
calc_theta(data, stat, quantile = NULL)
}
\arguments{
\item{data}{Data to calculate statistic on}

\item{stat}{Statistic to calculate on the data. Options: mean, median, sd, iqr, quantile}

\item{quantile}{Quantile to calculate between 0 and 1 if stat = quantile}
}
\value{
Value for the calculated statistic on the data
}
\description{
Calculates a specified statistic on a bootstrap sample
}
