% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_loc_stat.R
\name{get_pivotal_ci_loc}
\alias{get_pivotal_ci_loc}
\title{Computes the pivotal interval for a location-based statistic}
\usage{
get_pivotal_ci_loc(emp_dist, boot_dist, stat, quantile, alpha = 0.05)
}
\arguments{
\item{emp_dist}{Empirical distribution to bootstrap}

\item{boot_dist}{Bootstrap distribution for statistic}

\item{stat}{Measure of center to use as statistic. Options: mean, median, quantile}

\item{quantile}{Quantile to calculate between 0 and 1 if stat = quantile}

\item{alpha}{Alpha level for CI's. Default alpha = 0.05 for 95\% conf. Conf = (1 - alpha)*100\%}
}
\value{
Returns the pivotal CI for specified statistic
}
\description{
Computes the pivotal interval for a location-based statistic
}
