% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_scale_stat.R
\name{get_pivotal_ci_scale}
\alias{get_pivotal_ci_scale}
\title{Computes the pivotal interval for a scale-based statistic}
\usage{
get_pivotal_ci_scale(emp_dist, boot_dist, stat, alpha)
}
\arguments{
\item{emp_dist}{Empirical distribution to bootstrap}

\item{boot_dist}{Bootstrap distribution for statistic}

\item{stat}{Measure of variability to use as statistic. Options: sd, iqr}

\item{alpha}{Alpha level for CI's. Default alpha = 0.05 for 95\% conf. Conf = (1 - alpha)*100\%}
}
\value{
Returns the pivotal CI for specified statistic
}
\description{
Computes the pivotal interval for a scale-based statistic
}
